separate CacheValue is_valid / may_attempt / should_reattempt and StudIP specific functionality
