+++ studip_fuse/cache/__init__.py
@@ -1,5 +1,6 @@
 from studip_fuse.cache.async_cache import *
 from studip_fuse.cache.cached_session import CachedStudIPSession
 from studip_fuse.cache.circuit_breaker import *
+from studip_fuse.cache.studip_cache import *
 
-__all__ = async_cache.__all__ + circuit_breaker.__all__ + ["CachedStudIPSession"]
+__all__ = async_cache.__all__ + circuit_breaker.__all__ + ["CachedStudIPSession"] + studip_cache.__all__
